requirements.txt
requests
python-dotenv
anthropic
resend
PyGithub